Fix ChainDataProviderTest

+++ data/beaconrestapi/src/test/java/tech/pegasys/teku/beaconrestapi/handlers/v1/rewards/GetBlockRewardsTest.java
@@ -35,11 +35,13 @@ import tech.pegasys.teku.api.migrated.BlockRewardData;
 import tech.pegasys.teku.beaconrestapi.AbstractMigratedBeaconHandlerTest;
 import tech.pegasys.teku.infrastructure.async.SafeFuture;
 import tech.pegasys.teku.infrastructure.http.HttpStatusCodes;
+import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.spec.SpecMilestone;
 import tech.pegasys.teku.spec.datastructures.metadata.ObjectAndMetaData;
 
 public class GetBlockRewardsTest extends AbstractMigratedBeaconHandlerTest {
-  private final BlockRewardData data = new BlockRewardData(123, 283L, 672L, 982L, 198L);
+  private final BlockRewardData data =
+      new BlockRewardData(UInt64.valueOf(123), 283L, 672L, 982L, 198L);
   private final ObjectAndMetaData<BlockRewardData> blockRewardsResult =
       new ObjectAndMetaData<>(data, SpecMilestone.ALTAIR, false, true, true);
